requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.5
Django==4.1.2
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.5
screen==1.0.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
xlwt==1.3.0
